refactor: 경매 관련 페이지 폴더 구조 변경에 따른 url 변경 안된 부분 수정 및 api 추가 분리
