+++ apps/web/app/api/auction/bid/route.ts
@@ -1,6 +1,5 @@
+import { adminClient } from '@/app/admin';
 import { NextRequest, NextResponse } from 'next/server';
-
-import { adminClient } from '../../../admin';
 
 export async function POST(req: NextRequest) {
   const { auctionId, bidderId, bidPrice } = await req.json();
