+++ apps/web/app/api/auction/bidder-name/route.ts
@@ -1,6 +1,5 @@
+import { adminClient } from '@/app/admin';
 import { NextRequest, NextResponse } from 'next/server';
-
-import { adminClient } from '../../../admin';
 
 export async function GET(req: NextRequest) {
   const { searchParams } = new URL(req.url);
