+++ apps/web/app/api/auction/detail/[auctionId]/route.ts
@@ -1,6 +1,5 @@
+import { adminClient } from '@/app/admin';
 import { NextRequest, NextResponse } from 'next/server';
-
-import { adminClient } from '../../../../admin';
 
 // /api/auction/detail/[auctionId]
 export async function GET(req: NextRequest, { params }: { params: { auctionId: string } }) {
